Implement Withdraw method of WithdrawInterface using PostgreSQL

+++ internal/mocks/mock_store.go
@@ -183,7 +183,7 @@ func (m *MockWithdrawInterface) EXPECT() *MockWithdrawInterfaceMockRecorder {
 }
 
 // Withdraw mocks base method.
-func (m *MockWithdrawInterface) Withdraw(arg0 context.Context, arg1, arg2 string, arg3 int) (int, error) {
+func (m *MockWithdrawInterface) Withdraw(arg0 context.Context, arg1, arg2 string, arg3 float64) (int, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Withdraw", arg0, arg1, arg2, arg3)
 	ret0, _ := ret[0].(int)
